# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: Toolvio Backend API - RBAC & Dynamic Schema System
  version: 1.0.0
  description: |
    Comprehensive API documentation for Toolvio Backend with Role-Based Access Control (RBAC),
    Dynamic Schema Management, Audit Trail, and Multi-tenant Support.
    
    ## 🔐 Authentication
    Most endpoints require JWT authentication. Use the `/api/auth/login` endpoint to get a token,
    then include it in the Authorization header as `Bearer <token>`.
    
    ## 👥 User Roles
    - **Admin**: Full access to all features
    - **Office**: Can manage schemas and data, read audit logs
    - **Technician**: Can read schemas, manage data, read audit logs
    - **Customer**: Can only read data (no schema or audit access)
    
    ## 🏢 Multi-tenant
    All data is isolated by tenant. Users can only access data within their assigned tenant.
    
    Use the "Try it out" button to execute requests against your local server.
servers:
  - url: http://localhost:{port}
    description: Local server
    variables:
      port:
        default: "3000"
paths:
...
  /api/audit/{schemaName}/history:
    get:
      tags: [Audit]
      summary: Get schema audit history
      description: Retrieve audit history for a specific schema (requires audit read permission)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SchemaName'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Audit history retrieved
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
  /api/audit/{schemaName}/stats:
    get:
      tags: [Audit]
      summary: Get audit statistics
      description: Retrieve audit statistics for a schema (requires audit read permission)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SchemaName'
      responses:
        '200':
          description: Audit statistics retrieved
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
  /api/audit/{schemaName}/{recordId}/history:
    get:
      tags: [Audit]
      summary: Get document audit history
      description: Retrieve audit history for a specific document (requires audit read permission)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SchemaName'
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Document audit history retrieved
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
  /api/audit/{schemaName}/{recordId}/revert/{version}:
    post:
      tags: [Audit]
      summary: Revert document to version
      description: Revert a document to a specific version (requires audit rollback permission)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SchemaName'
        - $ref: '#/components/parameters/RecordId'
        - in: path
          name: version
          required: true
          schema:
            type: integer
          description: Version number to revert to
      responses:
        '200':
          description: Document reverted successfully
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions

  # System Routes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


      description: JWT token from /api/auth/login endpoint
  parameters:
    SchemaName:
      in: path
      name: schemaName
      required: true
      schema: { type: string }
      description: Dynamic schema name
...
    RecordId:
      in: path
      name: recordId
      required: true
      schema: { type: string }
      description: Record identifier
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...